Day 19: Beacon Scanner
This was a real tough one. I am not too big to admit I needed A LOT of help on solving this one. Thanks to a lot of topics on Reddit I was able to find the numerous flaws in my own logic.

I did not want to submit a solution until I fully understood what the code does and that is why the code is literally riddled with comments.