[2021] Day 22 - Optimization
After implementing part 2 I realized that it is probably much more efficient to apply that methodology to part 1 as well. That resulted in a reduction of execution time from almost 3 seconds to below 10ms.

Surprisingly it also reduced the execution time of part 2. Possibly due to either caching of less memory consumption.