[0000] Ignore all `*.txt` files from Git, except `example.txt`
In line with the rules of Advent of Code none of the puzzle inputs are to be distributed, but the examples are already publicly available, so those are okay, I guess.

Often puzzles have additional inputs for examples, and sometimes I create my own. Those should not be committed. By adding them to the `.gitignore` I prevent accidentally committing them.