[2023] Day 24 part 2
More than a year after the puzzle was started. I needed _a lot_ of help with this one. I hope that the explanation is good enough, also for myself if a similar puzzle happens again the future. Many kudos to all the people that post solutions and explanations on Reddit.