[2025] Some optimizations
I did not like the calculating of the duration, and I did not like it was copied for every project. I put this in a project `Shared` and updated all solution so far.

Some days have received optimized solutions I came across on Reddit. I left the original solutions in as well.